pyproject.toml
[project]
name = "shiftwise-be"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beanie<2.0",
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.7",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "lxml>=6.1.0",
    "motor>=3.7.1",
    "openai>=2.32.0",
    "pulp>=3.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.44.0",
]
